convObjectToPtr and convPtrToObject now utilize PointerObjects to keep the Java array and the C pointer to it together so the array can be released.
More work on serializing and deserializing Java objects of PTR_KIND to C 
pointers. 
CWrappers that have variable lengths when serialized should implement 
static Object newFromPointer(long) and void 
deserializeFromPointer(long). These methods are not added to the 
interface because they are only used from C and they are not needed if 
the serialized size is constant. Also, if the serialized size is 
constant after creation, only static Object newFromPointer(long) is 
required. Likewise, void deserializeFromPointer(long) is not necessary 
if the size is constant after creation.

No existing classes implement these functions. Existing code continues 
to function without it.