moved protocols and transports into their own xml files and renamed the jar files to indicate whether it is a protocol or a transport.
moved org.gnu.freeway.protocol.* to the main jar file so we don't need 
to include AbstractProtocol for every protocol jar

did the same with org.gnu.freeway.transport.*

moved MethodLister into SwitchTableGenerator and organized the imports

made tcp and udp attempt to use only Inet4Addresses. The actualy source 
of the buffer problems was IdentityService.

DynamicLibrary now longer looks for jar files in /usr/lib and 
/usr/local/lib. Why was it doing this to begin with? There are no jar 
files there and we *already had* the jar file that contains what we 
want to load.

CSMessage matches the C version more closely, but probably breaks a lot 
of stuff because some values are negative and there is more than one 0.