LRN: I'm tired of seeing something like "ATS tells us to switch to address '(null)' session 0x497198 for peer `IIH1' in state `S_NOT_CONNECTED'", so i wrote this.
- From what i see in the logs, '(null)'s are mostly caused by 0-byte
long addresses. I'm not sure whether it is a bug, or a feature (but
then, why would ATS tell us to switch to a 0-byte long address?).

CG: for inbound connections, ATS tells us to switch to a 0-byte long address
(since there is no address) but at the same time gives us a 'struct Session' 
to use as well.  So 0-byte addresses really mean 'inbound' connections.