LRN: Adding vectored exception handling for W32 (#1965)
Whenever an exception occurs, and the process is not being debugged, it will run a debugger specified by GNUNET_DEBUGGER environment variable, and wait for it to attach.
The net effect is the same as using JIT debugging (AeDebug), but without the stack being broken by SEH (because VEH has a priority over SEH), which allows for fuller backtraces for any exception, not just for GNUNET_abort() calls.