LRN:
Whenever a file is downloaded, store its downloaded name in a treestore
(FIXME: when a previously downloaded .gnd file is opened, set
downloaded name appropriately? Or not, since it won't have
anonymity?), as well as the anonymity level that was used.
When one of its children is downloaded, this information is checked
and, if present, is used to construct the child filename (and
anonymity is used to download the child; FIXME: anonymity == -1 might
be used, if parent supplied a filename, but not anonymity? This might
need a fix), so it is downloaded without showing the "Save as" dialog.
(TODO: the logic behind constructing the child filename from the name
of already downloaded directory file relies on directory file having a
'.gnd' extension; if it isn't there, the whole thing fails (tries to
mkdir() with the name of the .gnd-less directory file, which isn't
possible); Enforce .gnd extension? Change the name to avoid conflict?)

In future it might be necessary to add a context menu item that
overrides this behaviour and shows the dialog anyway.

Also changes the way gtkfilechooser is set up, so that it fills the
entry with short filename only, without the directory part, and the
directory part is used to select a directory. Without this the entry
gets absolute filename or directoryname, and if you later change the
directory in the dialog, contents of the entry remain the same. It's
intended to be relative most of the time!

Slight API changes for "Save as" dialog, now it invokes a callback
(FIXME: this breaks "Download URI", as it is not adapted to the
callback yet, that must be fixed as soon as possible!).
To that end one of its static functions was made public.